Add autocd
Add autocd so that you get the benefit of not having to type cd when you want to switch directories. i.e. you can type `..` instead of `cd ..` to move back a directory. Also will make it so you don't get the permission denied warning if you do this on accident with the way it is currently.